added typescript to Image.tsx
